make get_quids_dataset more flexible wrt. datatype
In my codebase, I use sometimes string-valued query IDs (so that I can debug things in my source data).
E.g. the constructor of this class works perfectly fine with string-valued query IDs.
This method expects query IDs to be integer-valued, which I think is an unnecessary restriction.

This change does not modify the default behavior, but allows for more flexibility.