use .sum() method to also support sparse matrices
When using a sparse matrix, e.g. scipy.sparse.csr_matrix, we otherwise get the error message: "NotImplementedError: adding a nonzero scalar to a sparse matrix is not supported".
With the .sum(), method, it works for both sparse and dense (numpy array) matrices.

This little change allows us to handle much bigger sparse datasets.
Memory saving depends on the dataset, I observed a factor of 7 for a dataset with about 5% density.